% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add-multi-filter-summaries.R
\name{add_filter_summaries}
\alias{add_filter_summaries}
\title{Add filter Summaries}
\usage{
add_filter_summaries(rowGroupDots, filterVars, data, summaryDots, otherFilters)
}
\arguments{
\item{rowGroupDots}{The parameter to be used as .dots in the group_by_ function.}

\item{filterVars}{A vector of the column names we want to generate summaries for. This column names are the vars of the filter we are generating the summaries for.}

\item{data}{The raw data subject of summarisation.}

\item{summaryDots}{The parameter to be used as .dots in the summarise_ function.}

\item{otherFilters}{A vector of the column names of every other collapsible filter.}
}
\value{
Returns a data set with several summarisations of the original data plus the original data.
Every filter is once ignored, not grouping by that filter thus generating totals
for that column which can then be retrieved by searching the label value "Samtliga".
}
\description{
Summarises data grouped by rowGroupDots while at the same time excluding the filterVars from the summary.
}

