% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add-multi-filter-summaries.R
\name{add_multi_filter_summaries}
\alias{add_multi_filter_summaries}
\title{Add multi filter Summaries}
\usage{
add_multi_filter_summaries(collapsibleFilters, rowGroupDots, data, summaryDots)
}
\arguments{
\item{collapsibleFilters}{A list of collapsible filters for which summaries should be generated}

\item{rowGroupDots}{The parameter to be used as .dots in the group_by_ function.}

\item{data}{The raw data subject of summarisation.}

\item{summaryDots}{The parameter to be used as .dots in the summarise_ function.}
}
\value{
Returns a data set with several summarisations of the original data plus the original data.
A filter being ignored means that the data will not be grouped by that filter thus generating subtotals not previously generated.
Every combination of ignored and not ignored filters is once processed.
}
\description{
Summarises data grouped by rowGroupDots while at the same time excluding the collapsibleFilters from the summary generating all permutations.
}

