% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preload-report.R
\name{convert_numerics_stored_as_char}
\alias{convert_numerics_stored_as_char}
\title{Convert all columns in a data frame that are convertible from character to numeric}
\usage{
convert_numerics_stored_as_char(data)
}
\arguments{
\item{data}{The data frame to be checked and modified accordingly}
}
\description{
Convert all columns in a data frame that are convertible from character to numeric
}

