% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/split-and-save-report-data.R
\name{create_double_multi_column_files}
\alias{create_double_multi_column_files}
\title{Create double multi column files}
\usage{
create_double_multi_column_files(actual, menuGroups, singleColumns,
  functionCalls, sexes, config, verbose, menuGroup, groupConfig,
  multipleColumnGroups, uniqueTimeValues)
}
\value{
The corresponding data sets are stored to disk. The data set is named
 "tableData" for immediate use in the Shiny frontend.
}
\description{
Generates all possible combinations of double multi column summaries of the baseData acording to the grouping and filtering parameters recieved.
Files are generated by using ready made multi column files.
}

