% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/split-and-save-report-data.R
\name{create_multi_column_files}
\alias{create_multi_column_files}
\title{Create multi column files}
\usage{
create_multi_column_files(config, metricNames, columnNames, sexNames, verbose,
  menuGroup, groupConfig, multipleColumnGroups, isDoubleMultiColumn,
  uniqueTimeValues)
}
\value{
The corresponding data sets are stored to disk. The data set is named
 "tableData" for immediate use in the Shiny frontend.
}
\description{
Generates all possible combinations of multi column summaries of the baseData acording to the grouping and filtering parameters recieved.
Files are generated by using ready made single column files.
}

