% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create-query-from-config.R
\name{create_query_from_config}
\alias{create_query_from_config}
\title{Get collapsible filters aplicable to a row}
\usage{
create_query_from_config(increment, config)
}
\arguments{
\item{increment}{A list containing the start and end parameters of the interval}

\item{config}{The config of the whole report}
}
\value{
Returns a list of filters applicable to a specific row as it is not possible to filter by the row itself or its childrows.
This filters are expressed as column names to filter by.
}
\description{
Filters collapsible filters in a config for a given row selection.
}

