% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/split-and-save-report-data.R
\name{create_single_column_files}
\alias{create_single_column_files}
\title{Create single column files}
\usage{
create_single_column_files(functionCalls, singleColumns, sexes, config,
  baseData, verbose, columnDefinitions)
}
\value{
The corresponding data sets are stored to disk. The data set is named
 "tableData" for immediate use in the Shiny frontend.
}
\description{
Generates all possible combinations of single column summaries of the baseData acording to the grouping and filtering parameters recieved.
}

