% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract-variables-names.R
\name{extract_variables_names}
\alias{extract_variables_names}
\title{Extract all Variable Names from Config}
\usage{
extract_variables_names(config, storageNames = c("var", "vars", "parent",
  "child"))
}
\arguments{
\item{config}{named list, config list to extract all variable names from}

\item{name}{character, name of the list item that holds varibales, default is \code{var}}
}
\description{
Extract all variable names from config and return a character vector. Can be used to create a SQL query.
}

