% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter-and-save-table-data.R
\name{filter_and_save_table_data_multicolumn}
\alias{filter_and_save_table_data_multicolumn}
\title{Filter and save table data multi column}
\usage{
filter_and_save_table_data_multicolumn(metric, column, sex, row, config,
  verbose, menuGroup, groupConfig, multipleColumnGroups, fileType = "base",
  isDoubleMultiColumn = FALSE, uniqueTimeValues)
}
\arguments{
\item{metric}{A string with the name of the metric for which the summary will be created}

\item{column}{A string with the name of the column for which the summary will be created}

\item{sex}{A string with the name of the sex for which the summary will be created}

\item{config}{The config of the whole report}

\item{verbose}{Verbose mode, i.e. tell the user what the program is doing}

\item{menuGroup}{The name of the group that is multi column and summaries will be merged by. i.e "metrics"}

\item{groupConfig}{The configuration of the menuGroup}

\item{multipleColumnGroups}{Name of the multi column group to merge by}

\item{fileType}{Defines which kind of file is to be generated. Options: base, child, subtotal}

\item{isDoubleMultiColumn}{A boolean that determines whether the summary to be created is multi column or double multi column.}

\item{rowgroup}{A string with the name of the row for which the summary will be created}
}
\value{
The corresponding data sets is stored to disk. The data set is named
 "tableData" for immediate use in the Shiny frontend.
}
\description{
Generates a multi column summary merging preexisting files acording to the grouping and filtering parameters recieved.
}

