% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter-and-save-table-data.R
\name{filter_and_save_table_data_single_column}
\alias{filter_and_save_table_data_single_column}
\title{Filter and save table data single column}
\usage{
filter_and_save_table_data_single_column(metric, column, sex, rowgroup, config,
  verbose, filteredBySexData, rowGroupDots, columnDefinitions, sexFilterDots,
  fileType = "base")
}
\arguments{
\item{metric}{A string with the name of the metric for which the summary will be created}

\item{column}{A string with the name of the column for which the summary will be created}

\item{sex}{A string with the name of the sex for which the summary will be created}

\item{rowgroup}{A string with the name of the row for which the summary will be created}

\item{config}{The config of the whole report}

\item{verbose}{Verbose mode, i.e. tell the user what the program is doing}

\item{filteredBySexData}{Data frame containing the raw data only for the sex passed in the sex parameter}

\item{rowGroupDots}{Vector of column names to group by in order to summarise.}

\item{columnDefinitions}{A list of column definitions (?)}

\item{sexFilterDots}{Message related. Remove?}

\item{fileType}{Defines which kind of file is to be generated. Options: base, child, subtotal}
}
\value{
The corresponding data sets is stored to disk. The data set is named
 "tableData" for immediate use in the Shiny frontend.
}
\description{
Generates a single column summary of the baseData acording to the grouping and filtering parameters recieved.
}

