% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-collapsible-filters-for-row.R
\name{get_collapsible_filters_for_row}
\alias{get_collapsible_filters_for_row}
\title{Get collapsible filters aplicable to a row}
\usage{
get_collapsible_filters_for_row(config, row)
}
\arguments{
\item{config}{The configuration of the report}

\item{row}{The row for which we want the available collapsible filters}
}
\value{
Returns a list of filters applicable to a specific row as it is not possible to filter by the row itself or its childrows.
This filters are expressed as column names to filter by.
}
\description{
Filters collapsible filters in a config for a given row selection.
}

