% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-double-multi-column-deliminiter.R
\name{get_double_multi_column_deliminiter}
\alias{get_double_multi_column_deliminiter}
\title{Get Double Multi Column Deliminiter}
\usage{
get_double_multi_column_deliminiter(asRegex = FALSE)
}
\arguments{
\item{asRegex}{logical, Should the return value be a regex for use in e.g. \code{\link{str_split}}?}
}
\description{
The names of double multi columns have two parts, which must be seprated in order to display the right header in the table. This is a helper function needed in the preloading and the server.R of the app.
}

