% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-list-of-desired-reports.R
\name{get_list_of_desired_reports}
\alias{get_list_of_desired_reports}
\title{Get list of desired reports}
\usage{
get_list_of_desired_reports(desiredTabs, listOfReports)
}
\arguments{
\item{desiredTabs}{A vector of tab names the scrip will be run for. If left empty the script will be run for all tabs.}

\item{listOfReports}{A list of reports to be filtered}
}
\description{
filters the listOfReports returning only those that fill the criteria of desiredTabs
}

