% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create-query-from-config.R
\name{get_min_max_string}
\alias{get_min_max_string}
\title{Get min and max string}
\usage{
get_min_max_string(increment, timeVariablesString)
}
\arguments{
\item{increment}{A list containing the start and end parameters of the interval}

\item{timeVariablesString}{A string containing the value of the time variable to be evaluated}
}
\value{
Returns a list of filters applicable to a specific row as it is not possible to filter by the row itself or its childrows.
This filters are expressed as column names to filter by.
}
\description{
Filters collapsible filters in a config for a given row selection.
}

