% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-time-variables.R
\name{get_time_variables}
\alias{get_time_variables}
\title{Get Time Variables}
\usage{
get_time_variables(config)
}
\arguments{
\item{config}{}
}
\description{
Get time variables form config
}

