% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-time-variables.R
\name{get_time_variables_alias}
\alias{get_time_variables_alias}
\title{Get Time Variables query}
\usage{
get_time_variables_alias(config)
}
\arguments{
\item{config}{}
}
\description{
Get time variables form config and return compound variable
}

