% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-list-of-desired-reports.R
\name{is_desired_path}
\alias{is_desired_path}
\title{Get list of desired reports}
\usage{
is_desired_path(path, desiredTabs)
}
\arguments{
\item{path}{A path to analyze}

\item{listOfReports}{A list of reports to be filtered}
}
\description{
filters the listOfReports returning only those that fill the criteria of desiredTabs
}

