% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_all_report_configs.R
\name{load_all_report_configs}
\alias{load_all_report_configs}
\title{Get a parseable list of all reports in application}
\usage{
load_all_report_configs(root)
}
\arguments{
\item{root}{The root directory to search for config files}
}
\value{
A list with the contents of all config files, sorted by folder structure
}
\description{
Get a parseable list of all reports in application
}

