% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load-cached-data.R
\name{load_cached_data}
\alias{load_cached_data}
\title{Load Cached Data}
\usage{
load_cached_data(reportConfig, cacheDirectory = "cache",
  envir = parent.frame(), verbose = FALSE)
}
\arguments{
\item{reportConfig}{the config for the report}

\item{cacheDir}{directory, where to look after the data file}

\item{...}{parameters passed to \code{\link{load}}}
}
\description{
Loads the cached data for a specific report
}

