% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pmspPreload.R
\docType{package}
\name{pmspPreload}
\alias{pmspPreload}
\alias{pmspPreload-package}
\title{pmspPreload: A package for preloading data for the Swedish Pensions Agency Statistics Portal}
\description{
pmspPreload: A package for preloading data for the Swedish Pensions Agency Statistics Portal
}

