% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preload-report.R
\name{preload_reports}
\alias{preload_reports}
\title{Preloading script for reports}
\usage{
preload_reports(sourceReportDir = file.path(getwd(), "modules/reports/"),
  targetAppDir = file.path(getwd(), "modules/reports/"), verbose = TRUE,
  logToFile = FALSE, cache = FALSE, cacheDirectory = "cache",
  desiredTabs = c(), increment = list(), scope = "all")
}
\arguments{
\item{sourceReportDir}{The source directory in which to search for report config files}

\item{targetAppDir}{The target app directory to which to write the pre-processed data and config files}

\item{verbose}{Verbose mode, i.e. tell the user what the program is doing}

\item{cache}{Store downloaded data locally; boolean. Note that activating local caching could take up a LOT of hard disk space!}

\item{cacheDirectory}{This is the location to store cached data. Required if \code{cache = TRUE}}

\item{desiredTabs}{A vector of tab names the scrip will be run for. If left empty the script will be run for all tabs.}

\item{increment}{A list that consists of two elements, from and to in order to set limits to the data to process.}

\item{scope}{Determines which parts of the scripts the user wants to run. Posible values 'logs', 'all'}
}
\description{
A script for preloading data and storing it in a pre-aggregated fashion for reports in the STA Shiny interface
}
\examples{
\dontrun{
preload_reports(cache = TRUE, logToFile = TRUE, desiredTabs = c("1-bt"))
}
}

