% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove-diplayName-from-config.R
\name{remove_display_name_from_config}
\alias{remove_display_name_from_config}
\title{Remove Display Name from Config Parts}
\usage{
remove_display_name_from_config(config)
}
\arguments{
\item{reportConfig}{}
}
\description{
Remove Display Name from Config Parts
}

