% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/split-and-save-report-data.R
\name{split_and_save_report_data}
\alias{split_and_save_report_data}
\title{Split and save report data}
\usage{
split_and_save_report_data(config, baseData, verbose = TRUE)
}
\value{
A number of data sets are stored to disk. All these data sets are named
 "tableData" for immediate use in the Shiny frontend.
}
\description{
Split a base dataset according to specifications in a config file
 and save it to a target directory
}
\examples{
\dontrun{
sourceReportDir = file.path(getwd(), "modules/reports/")
listOfAllReportTabs <- load_all_report_configs(sourceReportDir)
listOfReports <- listOfAllReportTabs[[1]]
config <- listOfReports$reports[[1]]
baseData <- fetch_some_raw_data_from_sql_server() # Mock code

split_and_save_report_data(config, baseData)
}
}

