% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add-multi-filter-summaries.R
\name{summarise_removing_filters}
\alias{summarise_removing_filters}
\title{Summarise removing filters}
\usage{
summarise_removing_filters(rowGroupDots, removedFilterVars, data, summaryDots)
}
\arguments{
\item{rowGroupDots}{The parameter to be used as .dots in the group_by_ function.}

\item{removedFilterVars}{A vector of names of filtered colums we want to avoid grouping by.}

\item{data}{The raw data subject of summarisation.}

\item{summaryDots}{The parameter to be used as .dots in the summarise_ function.}
}
\value{
Returns a data set with several summarisations of the original data.
Every removedFilter is once ignored, not grouping by that filter thus generating totals
for that column which can then be retrieved by searching the label value "Samtliga".
}
\description{
Summarises data grouped by rowGroupDots while at the same time excluding the removedFilters from the summary
}

