% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/verify-or-create-target-file-structure.R
\name{verify_or_create_target_file_structure}
\alias{verify_or_create_target_file_structure}
\title{Verify or create the necessary file structure for a report.}
\usage{
verify_or_create_target_file_structure(config)
}
\arguments{
\item{config}{A config list, containing the 'target' field}
}
\description{
Verify or create the necessary file structure for a report.
}

